# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: kocmoc messenger (legacy Node.js)
  version: 1.0.0
  description: >-
    Снимок текущих REST маршрутов. Ответы используют поля `error` для ошибок и
    `message`/`token`/`success` для успешных операций. JWT срок 30 дней.
servers:
  - url: http://localhost:3000
paths:
...
  /api/files/{fileId}:
    get:
      summary: Получить файл
      parameters:
        - in: path
          name: fileId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Бинарный поток
          content:
            '*/*': { schema: { type: string, format: binary } }
        '404': { $ref: '#/components/responses/Error' }
    delete:
      summary: Удалить файл
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Message' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Ошибка
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
    Message:
      description: Ответ с текстом
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string }
...